07-RESTful API及requests/RestDemo/requirements.txt
django=2.2.10
djangorestframework==3.12.1
django-filter=2.4.0
django-rest-swagger==2.2.0